domd: force enable UFS
This is adopted commit
248a93b6198aed8b3ea121ea7c68152d0bc610ca
"[gen4] domd: force enable UFS"
from aoscloud/meta-aos-rcar-demo2023
with changes for domd-spyder.cfg already in aos-core-zephyr.

UFS was disabled by Renesas till there is a correct support for Clock
IC initialization in IPL. But we need a second storage in our
use-case. So we are enabling UFS back.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Signed-off-by: Sergiy Kibrik <sergiy_kibrik@epam.com>
Acked-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>